fix(auth.controller.ts): set sameSite cookie attribute to "none"
The sameSite attribute was set to "strict", which prevents the cookie
from being sent in cross-site requests. This caused issues when the
frontend and backend were running on different domains.

Setting the sameSite attribute to "none" allows the cookie to be sent
in cross-site requests, but it also requires the secure attribute to
be set to true. This ensures that the cookie is only sent over HTTPS.